moved login and home into pages
